+++ app/data/repositories/layer0/repository.py
@@ -1,12 +1,8 @@
-import json
-
 import structlog
 
-from app.data import model, template
-from app.data.repositories.layer0 import tables
-from app.data.repositories.layer0.common import RAWDATA_SCHEMA
+from app.data import model
+from app.data.repositories.layer0 import objects, old_objects, tables
 from app.lib.storage import enums, postgres
-from app.lib.web.errors import DatabaseError
 
 
 class Layer0Repository(postgres.TransactionalPGRepository):
@@ -15,6 +11,8 @@ class Layer0Repository(postgres.TransactionalPGRepository):
         super().__init__(storage)
 
         self.table_repo = tables.Layer0TableRepository(storage)
+        self.objects_repo = objects.Layer0ObjectRepository(storage)
+        self.old_objects_repo = old_objects.Layer0OldObjectsRepository(storage)
 
     def create_table(self, data: model.Layer0TableMeta) -> model.Layer0CreationResponse:
         return self.table_repo.create_table(data)
@@ -44,122 +42,23 @@ class Layer0Repository(postgres.TransactionalPGRepository):
         table_id: int,
         objects: list[model.Layer0Object],
     ) -> None:
-        query = "INSERT INTO rawdata.objects (id, table_id, data) VALUES "
-        params = []
-        values = []
-
-        for obj in objects:
-            values.append("(%s, %s, %s)")
-            params.extend([obj.object_id, table_id, json.dumps(obj.data, cls=model.CatalogObjectEncoder)])
-
-        query += ",".join(values)
-        query += " ON CONFLICT (id) DO UPDATE SET data = EXCLUDED.data, table_id = EXCLUDED.table_id"
-
-        self._storage.exec(query, params=params)
+        return self.objects_repo.upsert_objects(table_id, objects)
 
     def upsert_old_object(
         self,
         table_id: int,
         processing_info: model.Layer0OldObject,
     ) -> None:
-        self._storage.exec(
-            """
-            INSERT INTO rawdata.old_objects (table_id, object_id, status, data, metadata)
-            VALUES (%s, %s, %s, %s, %s) 
-            ON CONFLICT (table_id, object_id) DO 
-                UPDATE SET status = EXCLUDED.status, metadata = EXCLUDED.metadata
-            """,
-            params=[
-                table_id,
-                processing_info.object_id,
-                processing_info.status,
-                json.dumps(processing_info.data, cls=model.CatalogObjectEncoder),
-                json.dumps(processing_info.metadata),
-            ],
-        )
+        return self.old_objects_repo.upsert_old_object(table_id, processing_info)
 
     def get_table_statistics(self, table_id: int) -> model.TableStatistics:
-        statuses_query = """
-            SELECT COALESCE(status, 'unprocessed') AS status, COUNT(1) 
-            FROM rawdata.processing AS p
-            RIGHT JOIN rawdata.objects AS o ON p.object_id = o.id
-            WHERE o.table_id = %s
-            GROUP BY status"""
-
-        stats_query = """
-            SELECT MAX(modification_dt) AS modification_dt , COUNT(1) AS cnt
-            FROM rawdata.objects 
-            WHERE table_id = %s"""
-
-        status_rows = self._storage.query(statuses_query, params=[table_id])
-        stats = self._storage.query_one(stats_query, params=[table_id])
-
-        table_name = self._storage.query_one(template.GET_RAWDATA_TABLE, params=[table_id]).get("table_name")
-        if table_name is None:
-            raise DatabaseError(f"unable to fetch table with id {table_id}")
-
-        total_original_rows_query = f'SELECT COUNT(1) AS cnt FROM {RAWDATA_SCHEMA}."{table_name}"'
-
-        total_original_rows = self._storage.query_one(total_original_rows_query).get("cnt")
-        if total_original_rows is None:
-            raise DatabaseError(f"unable to fetch total rows for table {table_name}")
-
-        return model.TableStatistics(
-            {enums.ObjectProcessingStatus(row["status"]): row["count"] for row in status_rows},
-            stats["modification_dt"],
-            stats["cnt"],
-            total_original_rows,
-        )
+        return self.objects_repo.get_table_statistics(table_id)
 
     def get_old_object_statuses(self, table_id: int) -> dict[enums.ObjectProcessingStatus, int]:
-        rows = self._storage.query(
-            """
-            SELECT status, COUNT(*) FROM rawdata.old_objects 
-            WHERE table_id = %s 
-            GROUP BY status
-            """,
-            params=[table_id],
-        )
-
-        return {enums.ObjectProcessingStatus(row["status"]): row["count"] for row in rows}
+        return self.old_objects_repo.get_old_object_statuses(table_id)
 
     def get_objects(self, table_id: int, limit: int, offset: int) -> list[model.Layer0Object]:
-        rows = self._storage.query(
-            """
-            SELECT id, data
-            FROM rawdata.objects
-            WHERE table_id = %s
-            LIMIT %s OFFSET %s
-            """,
-            params=[table_id, limit, offset],
-        )
-
-        return [
-            model.Layer0Object(
-                row["id"],
-                json.loads(json.dumps(row["data"]), cls=model.CatalogObjectDecoder),
-            )
-            for row in rows
-        ]
+        return self.objects_repo.get_objects(table_id, limit, offset)
 
     def get_old_objects(self, table_id: int, batch_size: int, offset: int) -> list[model.Layer0OldObject]:
-        rows = self._storage.query(
-            """
-            SELECT object_id, pgc, status, data, metadata
-            FROM rawdata.old_objects
-            WHERE table_id = %s
-            LIMIT %s OFFSET %s
-            """,
-            params=[table_id, batch_size, offset],
-        )
-
-        return [
-            model.Layer0OldObject(
-                row["object_id"],
-                enums.ObjectProcessingStatus(row["status"]),
-                row["metadata"],
-                json.loads(json.dumps(row["data"]), cls=model.CatalogObjectDecoder),
-                row["pgc"],
-            )
-            for row in rows
-        ]
+        return self.old_objects_repo.get_old_objects(table_id, batch_size, offset)
